changed payload_size to uint16_t

+++ production/inc/internal/db/gaia_se_object.hpp
@@ -21,9 +21,10 @@ namespace db
 struct gaia_se_object_t {
     gaia_id_t id;
     gaia_type_t type;
-    // The Flatbuffer size limit is 2GB (signed 32-bit). Total size of the payload will
-    // be the serialized flatbuffer size plus the num_references * sizeof(gaia_id_t).
-    uint32_t payload_size;
+    // The Flatbuffer size limit is 2GB (signed 32-bit). With a 16-bit payload size,
+    // the limit is 65,536 bytes. This total size of the payload will be the
+    // serialized flatbuffer size plus the num_references * sizeof(gaia_id_t).
+    uint16_t payload_size;
     uint16_t num_references;
     char payload[0];
 };
